+++ src/components/card/MissionReportCard.tsx
@@ -103,7 +103,7 @@ const MissionReportCard: React.FC<MissionReportProps> = ({
       >
         {type === 'MISSION_POINT' ? (
           <span>
-            จำนวนแต็มที่ได้รับ :{' '}
+            จำนวนแต้มที่ได้รับ :{' '}
             {detailReward?.point ? formatNumberWithCommas(parseFloat(detailReward?.point)) : 0} แต้ม
           </span>
         ) : (